refact models

+++ store/cart/models.py
@@ -5,10 +5,28 @@ from shop.models import Product
 
 
 class Cart(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE, verbose_name='Пользователь')
-    total_price = models.PositiveIntegerField(default=0, verbose_name='Общая стоимость')
-    time_created = models.DateTimeField(auto_now_add=True, verbose_name='Время создания')
-    time_updated = models.DateTimeField(auto_now=True, verbose_name='Время изменения')
+    user = models.ForeignKey(
+        User,
+        on_delete=models.CASCADE,
+        verbose_name='Пользователь'
+    )
+
+    total_price = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        default=0,
+        verbose_name='Общая стоимость'
+    )
+
+    time_created = models.DateTimeField(
+        auto_now_add=True,
+        verbose_name='Время создания'
+    )
+
+    time_updated = models.DateTimeField(
+        auto_now=True,
+        verbose_name='Время изменения'
+    )
 
     class Meta:
         ordering = ('time_created',)
@@ -20,14 +38,41 @@ class Cart(models.Model):
 
 
 class CartProduct(models.Model):
-    product = models.ForeignKey(Product, related_name='cart_products',
-                                on_delete=models.CASCADE, verbose_name='Товар')
-    quantity = models.PositiveIntegerField(default=1, verbose_name='Количество')
-    several_price = models.PositiveIntegerField(default=0, verbose_name='Цена за количество')
-    time_created = models.DateTimeField(auto_now_add=True, verbose_name='Время создания')
-    time_updated = models.DateTimeField(auto_now=True, verbose_name='Время изменения')
-    cart = models.ForeignKey('Cart', related_name='cart_products',
-                             on_delete=models.CASCADE, verbose_name='Корзина')
+    product = models.ForeignKey(
+        Product,
+        related_name='cart_products',
+        on_delete=models.CASCADE,
+        verbose_name='Товар'
+    )
+
+    quantity = models.PositiveIntegerField(
+        default=1,
+        verbose_name='Количество'
+    )
+
+    several_price = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        default=0,
+        verbose_name='Цена за количество'
+    )
+
+    time_created = models.DateTimeField(
+        auto_now_add=True,
+        verbose_name='Время создания'
+    )
+
+    time_updated = models.DateTimeField(
+        auto_now=True,
+        verbose_name='Время изменения'
+    )
+
+    cart = models.ForeignKey(
+        'Cart',
+        related_name='cart_products',
+        on_delete=models.CASCADE,
+        verbose_name='Корзина'
+    )
 
     class Meta:
         ordering = ('time_created',)
